@@ -69,52 +69,41 @@
         <app.version>${project.version}</app.version>
         <commons-collections4.version>4.4</commons-collections4.version>
         <commons-io.version>2.8.0</commons-io.version>
-        <java.version>11</java.version>
-        <jdk.source>11</jdk.source>
-        <jdk.target>11</jdk.target>
+        <java.version>17</java.version>
+        <jdk.source>17</jdk.source>
+        <jdk.target>17</jdk.target>
         <google-api-services-admin-directory.version>directory_v1-rev118-1.25.0</google-api-services-admin-directory.version>
         <google-api-services-drive.version>v3-rev20200813-1.30.10</google-api-services-drive.version>
         <google-api-services-groupssettings.version>v1-rev20200617-1.30.10</google-api-services-groupssettings.version>
         <google-auth-library-oauth2-http.version>0.21.1</google-auth-library-oauth2-http.version>
         <jquery.version>3.5.1</jquery.version>
-        <lms-canvas-services.version>5.0.10</lms-canvas-services.version>
-        <lms-canvas-common-configuration.version>5.0.7</lms-canvas-common-configuration.version>
-        <lms-canvas-iu-custom-services.version>5.0.11</lms-canvas-iu-custom-services.version>
-        <lms-canvas-lti-framework.version>5.1.15</lms-canvas-lti-framework.version>
         <lms-canvas-rivet.version>5.2.2.0</lms-canvas-rivet.version>
-        <lms-canvas-redis-config.version>5.0.7</lms-canvas-redis-config.version>
-        <lms-email-service.version>5.0.6</lms-email-service.version>
-        <nimbus-jose-jwt.version>9.10.1</nimbus-jose-jwt.version>
-        <spring-cloud-starter-vault-config.version>3.0.5</spring-cloud-starter-vault-config.version>
-        <springdoc-openapi-ui.version>1.6.9</springdoc-openapi-ui.version>
-        <spring-security-bom.version>5.5.2</spring-security-bom.version>
+        <lms-embedded-services.version>5.2.5</lms-embedded-services.version>
+        <lms-team-spring-boot-it12>4.8</lms-team-spring-boot-it12>
+        <springdoc-openapi-ui.version>1.6.15</springdoc-openapi-ui.version>
+        <spring-cloud-starter-parent.version>2021.0.5</spring-cloud-starter-parent.version>
         <webjars-fontawesome.version>4.7.0</webjars-fontawesome.version>
-        <webjars-locator.version>0.42</webjars-locator.version>
+        <webjars-locator.version>0.46</webjars-locator.version>
 
         <git-commit-id-maven-plugin.version>5.0.0</git-commit-id-maven-plugin.version>
-        <hibernate5-ddl-maven-plugin.version>2.3.0</hibernate5-ddl-maven-plugin.version>
-        <plugins.compiler.version>3.8.1</plugins.compiler.version>
-        <plugins.gpg.version>1.6</plugins.gpg.version>
-        <plugins.javadoc.version>3.0.1</plugins.javadoc.version>
+        <hibernate5-ddl-maven-plugin.version>2.5.0</hibernate5-ddl-maven-plugin.version>
+        <plugins.compiler.version>3.10.1</plugins.compiler.version>
+        <plugins.gpg.version>3.0.1</plugins.gpg.version>
+        <plugins.javadoc.version>3.5.0</plugins.javadoc.version>
         <plugins.license.version>2.0.0</plugins.license.version>
-        <plugins.nexus-staging.version>1.6.8</plugins.nexus-staging.version>
+        <plugins.nexus-staging.version>1.6.13</plugins.nexus-staging.version>
         <plugins.release.version>2.5.3</plugins.release.version>
-        <plugins.source.version>3.0.1</plugins.source.version>
+        <plugins.source.version>3.2.1</plugins.source.version>
     </properties>
 
     <dependencyManagement>
         <dependencies>
             <dependency>
-                <groupId>org.springframework.security</groupId>
-                <artifactId>spring-security-bom</artifactId>
-                <version>${spring-security-bom.version}</version>
+                <groupId>org.springframework.cloud</groupId>
+                <artifactId>spring-cloud-starter-parent</artifactId>
+                <version>${spring-cloud-starter-parent.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
-            </dependency>
-            <dependency>
-                <groupId>com.nimbusds</groupId>
-                <artifactId>nimbus-jose-jwt</artifactId>
-                <version>${nimbus-jose-jwt.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
