@@ -8,18 +8,18 @@ package edu.iu.uits.lms.gct.model;
  * %%
  * Redistribution and use in source and binary forms, with or without modification,
  * are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice, this
  *    list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the Indiana University nor the names of its contributors
  *    may be used to endorse or promote products derived from this software without
  *    specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
@@ -54,8 +54,8 @@ public class DecoratedCanvasUser {
       this.email = user.getEmail();
       this.loginId = user.getLoginId();
       this.enrollments = user.getEnrollments().stream()
-            .map(DecoratedEnrollment::new)
-            .collect(Collectors.toList());
+              .map(DecoratedEnrollment::new)
+              .collect(Collectors.toList());
    }
 
    public boolean isTeacher() {
