@@ -50,6 +50,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.metrics.buffering.BufferingApplicationStartup;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 
@@ -72,8 +73,14 @@ public class WebApplication {
     @Autowired
     private ToolConfig toolConfig;
 
+    private final static int STEP_CAPACITY = 2048;
+
     public static void main(String[] args) {
-        SpringApplication.run(WebApplication.class, args);
+        SpringApplication app = new SpringApplication(WebApplication.class);
+        BufferingApplicationStartup startup = new BufferingApplicationStartup(STEP_CAPACITY);
+        startup.addFilter(startupStep -> startupStep.getName().matches("spring.boot.application.ready"));
+        app.setApplicationStartup(startup);
+        app.run(args);
     }
 
     @Autowired
