@@ -33,12 +33,15 @@ package edu.iu.uits.lms.gct.config;
  * #L%
  */
 
+import edu.iu.uits.lms.common.it12logging.LmsFilterSecurityInterceptorObjectPostProcessor;
+import edu.iu.uits.lms.common.it12logging.RestSecurityLoggingConfig;
 import edu.iu.uits.lms.common.oauth.CustomJwtAuthenticationConverter;
 import edu.iu.uits.lms.lti.repository.DefaultInstructorRoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.security.SecurityProperties;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
+import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
@@ -60,6 +63,7 @@ public class SecurityConfig {
             http.requestMatchers().antMatchers("/rest/**", "/api/**")
                   .and()
                   .authorizeRequests()
+                  .antMatchers(HttpMethod.OPTIONS, "/rest/**").permitAll()
                   .antMatchers("/rest/**")
                   .access("hasAuthority('SCOPE_lms:rest') and hasAuthority('ROLE_LMS_REST_ADMINS')")
                   .antMatchers("/api/**").permitAll()
@@ -68,6 +72,8 @@ public class SecurityConfig {
                   .and()
                   .oauth2ResourceServer()
                   .jwt().jwtAuthenticationConverter(new CustomJwtAuthenticationConverter());
+
+            http.apply(new RestSecurityLoggingConfig());
         }
     }
 
@@ -86,7 +92,8 @@ public class SecurityConfig {
                   .and()
                   .authorizeRequests()
                   .antMatchers(WELL_KNOWN_ALL, "/error").permitAll()
-                  .antMatchers("/**").hasRole(BASE_USER_ROLE);
+                  .antMatchers("/**").hasRole(BASE_USER_ROLE)
+                  .withObjectPostProcessor(new LmsFilterSecurityInterceptorObjectPostProcessor());
 
             //Setup the LTI handshake
             Lti13Configurer lti13Configurer = new Lti13Configurer()
@@ -98,7 +105,8 @@ public class SecurityConfig {
             http.requestMatchers().antMatchers("/**")
                   .and()
                   .authorizeRequests()
-                  .anyRequest().authenticated();
+                  .anyRequest().authenticated()
+                  .withObjectPostProcessor(new LmsFilterSecurityInterceptorObjectPostProcessor());
 
             http.exceptionHandling().accessDeniedPage("/accessDenied");
         }
