@@ -39,6 +39,7 @@ import io.swagger.v3.oas.annotations.info.Info;
 import io.swagger.v3.oas.annotations.security.OAuthFlow;
 import io.swagger.v3.oas.annotations.security.OAuthFlows;
 import io.swagger.v3.oas.annotations.security.OAuthScope;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.security.SecurityScheme;
 import org.springdoc.core.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
@@ -47,7 +48,8 @@ import org.springframework.context.annotation.Profile;
 
 @Profile("swagger")
 @Configuration
-@OpenAPIDefinition(info = @Info(title = "Google Course Tools REST Endpoints", version = "${gct.version}"))
+@OpenAPIDefinition(info = @Info(title = "Google Course Tools REST Endpoints", version = "${gct.version}"),
+      security = @SecurityRequirement(name = "security_auth_gct"))
 @SecurityScheme(name = "security_auth_gct", type = SecuritySchemeType.OAUTH2,
       flows = @OAuthFlows(authorizationCode = @OAuthFlow(
             authorizationUrl = "${springdoc.oAuthFlow.authorizationUrl}",
